More refactoring and unit-testing. Make ChefMaker now be an instance of a Switch Remove the power switch from other devices
